pyproject.toml
[tool.poetry]
name = "greptimedb-sqlparse-fault"
version = "0.1.0"
description = ""
authors = ["Ildar Akhmetgaleev <akhilman@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mysql-connector-python = "^9.0.0"
mysqlclient = "^2.2.4"
click = "^8.1.7"


[tool.poetry.group.dev.dependencies]
wheel = "^0.44.0"
ipdb = "^0.13.13"
ipython = "^8.27.0"
python-lsp-server = {extras = ["rope"], version = "^1.12.0"}
pytype = "^2024.4.11"
ruff = "^0.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
